+++ main.go
@@ -8,6 +8,7 @@ import (
 	"immich-go/cmdduplicate"
 	"immich-go/cmdmetadata"
 	"immich-go/cmdstack"
+	"immich-go/cmdtool"
 	"immich-go/cmdupload"
 	"immich-go/immich"
 	"immich-go/immich/logger"
@@ -142,6 +143,8 @@ func Run(ctx context.Context, log *logger.Logger) (*logger.Logger, error) {
 		err = cmdmetadata.MetadataCommand(ctx, app.Immich, app.Logger, flag.Args()[1:])
 	case "stack":
 		err = cmdstack.NewStackCommand(ctx, app.Immich, app.Logger, flag.Args()[1:])
+	case "tool":
+		err = cmdtool.CommandTool(ctx, app.Immich, app.Logger, flag.Args()[1:])
 	default:
 		err = fmt.Errorf("unknwon command: %q", cmd)
 	}