+++ cmdupload/upload_test.go
@@ -35,6 +35,9 @@ func (c *stubIC) AddAssetToAlbum(context.Context, string, []string) ([]immich.Up
 func (c *stubIC) CreateAlbum(context.Context, string, []string) (immich.AlbumSimplified, error) {
 	return immich.AlbumSimplified{}, nil
 }
+func (c *stubIC) UpdateAssets(ctx context.Context, IDs []string, isArchived bool, isFavorite bool, removeParent bool, stackParentId string) error {
+	return nil
+}
 
 // type mockedBrowser struct {
 // 	assets []assets.LocalAssetFile